requirements.txt
streamlit==1.44.1
langgraph==0.3.20
langchain==0.3.21
fpdf==1.7.2
Pillow==11.1.0
python-pptx==1.0.2
litellm==1.64.1
python-dotenv==1.1.0
google-generativeai==0.8.4 